feat: renamed ArtistTrackList to TrackList and cleaned up code
