+++ app/(music)/artist/[id]/ArtistView.tsx
@@ -1,7 +1,6 @@
 import FollowButton from "@/components/FollowButton";
 import AudioFeatures from "@/components/music/AudioFeatures";
-import { AlbumList } from "@/components/music/lists";
-import ArtistTrackList from "@/components/music/lists/ArtistTrackList";
+import { AlbumList, TrackList } from "@/components/music/lists";
 import MusicHeader from "@/components/music/MusicHeader";
 import PlayContext from "@/components/music/PlayContext";
 import { formatNumber } from "@/lib/formatters";
@@ -46,7 +45,7 @@ const ArtistView = ({
     <AudioFeatures audioFeatures={averageAudioFeatures} />
     <div className="divider" />
     <h4 className="font-black uppercase">Top Tracks</h4>
-    <ArtistTrackList tracks={topTracks.tracks} showAlbum />
+    <TrackList tracks={topTracks.tracks} showAlbum />
     <h4 className="font-black uppercase">Recent Albums</h4>
     <AlbumList albums={albums.items} />
   </>
