+++ components/music/lists/index.tsx
@@ -1,2 +1,3 @@
 export { default as AlbumList } from "./AlbumList";
+export { default as ContextTrackList } from "./ContextTrackList";
 export { default as TrackList } from "./TrackList";
